extract the "core APIs" into a rayon-core crate
The parallel iterator APIs remain in the rayon crate. This allows the
rayon crate to advance a major version without introducing two thread-pools.
