+++ rayon-core/src/configuration.rs
