+++ rayon-core/src/future/test.rs
