+++ rayon-core/src/latch.rs
