+++ rayon-core/src/log.rs
