+++ rayon-core/src/registry.rs
