+++ rayon-core/src/scope/test.rs
@@ -3,7 +3,6 @@ extern crate rand;
 use Configuration;
 use {scope, Scope};
 use ThreadPool;
-use prelude::*;
 use rand::{Rng, SeedableRng, XorShiftRng};
 use std::cmp;
 use std::iter::once;
@@ -70,27 +69,6 @@ fn divide_and_conquer_seq(counter: &AtomicUsize, size: usize) {
         // count the leaves
         counter.fetch_add(1, Ordering::SeqCst);
     }
-}
-
-#[test]
-fn scope_mix() {
-    let counter_p = &AtomicUsize::new(0);
-    scope(|s| {
-        s.spawn(move |s| {
-            divide_and_conquer(s, counter_p, 1024);
-        });
-        s.spawn(move |_| {
-            let a: Vec<i32> = (0..1024).collect();
-            let r1 = a.par_iter()
-                .weight_max()
-                .map(|&i| i + 1)
-                .reduce_with(|i, j| i + j);
-            let r2 = a.iter()
-                .map(|&i| i + 1)
-                .fold(0, |a, b| a + b);
-            assert_eq!(r1.unwrap(), r2);
-        });
-    });
 }
 
 struct Tree<T> {
