+++ rayon-core/src/sleep/mod.rs
