+++ rayon-core/src/util.rs
