+++ src/lib.rs
@@ -7,47 +7,29 @@
 #![cfg_attr(not(feature = "unstable"), allow(warnings))]
 
 extern crate deque;
-#[macro_use]
-extern crate lazy_static;
 #[cfg(feature = "unstable")]
 extern crate futures;
 extern crate libc;
 extern crate num_cpus;
 extern crate rand;
+extern crate rayon_core;
 
-#[macro_use]
-mod log;
-
-mod configuration;
-mod latch;
-mod join;
-mod job;
 pub mod par_iter;
 pub mod prelude;
-#[cfg(test)]
 mod test;
-mod registry;
-#[cfg(feature = "unstable")]
-mod future;
-mod scope;
-mod sleep;
-#[cfg(feature = "unstable")]
-mod spawn_async;
-mod thread_pool;
-mod unwind;
-mod util;
 
-pub use configuration::Configuration;
-pub use configuration::PanicHandler;
-pub use configuration::InitError;
-pub use configuration::dump_stats;
-pub use configuration::initialize;
-pub use thread_pool::ThreadPool;
-pub use join::join;
-pub use scope::{scope, Scope};
+pub use rayon_core::current_num_threads;
+pub use rayon_core::Configuration;
+pub use rayon_core::PanicHandler;
+pub use rayon_core::InitError;
+pub use rayon_core::dump_stats;
+pub use rayon_core::initialize;
+pub use rayon_core::ThreadPool;
+pub use rayon_core::join;
+pub use rayon_core::{scope, Scope};
 #[cfg(feature = "unstable")]
-pub use spawn_async::spawn_async;
+pub use rayon_core::spawn_async;
 #[cfg(feature = "unstable")]
-pub use spawn_async::spawn_future_async;
+pub use rayon_core::spawn_future_async;
 #[cfg(feature = "unstable")]
-pub use future::RayonFuture;
+pub use rayon_core::RayonFuture;
