+++ src/par_iter/internal.rs
@@ -3,10 +3,9 @@
 //! parallel iterators should not need to interact with them directly.
 //! See `README.md` for a high-level overview.
 
-use join;
+use rayon_core::join;
 use super::IndexedParallelIterator;
 use super::len::*;
-use registry::Registry;
 
 pub trait ProducerCallback<ITEM> {
     type Output;
@@ -162,7 +161,7 @@ impl Splitter {
 
     #[inline]
     fn new_thief() -> Splitter {
-        Splitter::Thief(Splitter::thief_id(), Registry::current_num_threads())
+        Splitter::Thief(Splitter::thief_id(), ::current_num_threads())
     }
 
     #[inline]
@@ -181,7 +180,7 @@ impl Splitter {
                 let id = Splitter::thief_id();
                 if *origin != id {
                     *origin = id;
-                    *splits = Registry::current_num_threads();
+                    *splits = ::current_num_threads();
                     true
                 } else if *splits > 0 {
                     *splits /= 2;
